chore: update test coverrage
