+++ cmd/create.go
@@ -2,12 +2,7 @@ package cmd
 
 import (
 	"fmt"
-	"log"
-	"os"
-	"time"
 
-	"github.com/aws/aws-sdk-go/service/s3"
-	"github.com/orange-cloudavenue/cloudavenue-sdk-go/v1/infrapi"
 	"github.com/spf13/cobra"
 )
 
@@ -74,245 +69,4 @@ func init() {
 		fmt.Println("Error from Flag Name, is require.", err)
 		return
 	}
-}
-
-// createPublicIPCmd create a public ip resource(s)
-var createPublicIPCmd = &cobra.Command{
-	Use:               argPublicIP,
-	Short:             "Create an ip",
-	Example:           "ip create --name <EdgeGateway>",
-	DisableAutoGenTag: true,
-	Run: func(cmd *cobra.Command, args []string) {
-		// init Config File
-		v, err := initConfig()
-		if err != nil {
-			log.Default().Println("Error from init config file", err)
-			return
-		}
-
-		// init Client
-		err = initClient(v)
-		if err != nil {
-			log.Default().Println("Error from init client:", err)
-			os.Exit(1)
-		}
-
-		// Check if time flag is set
-		if cmd.Flag(flagTime).Value.String() == "true" {
-			defer timeTrack(time.Now(), cmd.CommandPath())
-		}
-
-		// Get the name from the command line
-		gwName, err := cmd.Flags().GetString(flagName)
-		if err != nil {
-			fmt.Println("Malformed argument EdgeGateway Name ", err)
-			return
-		}
-
-		// Create a public ip
-		s.Stop()
-		fmt.Println("create public ip resource")
-		fmt.Println("for EdgeGateway name: " + gwName)
-		s.Restart()
-
-		job, err := c.V1.PublicIP.New(gwName)
-		if err != nil {
-			fmt.Println("Unable to create public ip", err)
-			return
-		}
-		err = job.Wait(5, 300)
-		if err != nil {
-			fmt.Println("Error during public ip creation !!", err)
-			return
-		}
-		s.FinalMSG = "public ip resource created successfully !!"
-		s.Stop()
-	},
-}
-
-// createVDCCmd create a vdc resource(s)
-var createVDCCmd = &cobra.Command{
-	Use:               argVDC,
-	Short:             "Create an vdc",
-	Example:           "vdc create --name <vdc name>",
-	DisableAutoGenTag: true,
-	Run: func(cmd *cobra.Command, args []string) {
-		// init Config File
-		v, err := initConfig()
-		if err != nil {
-			log.Default().Println("Error from init config file", err)
-			return
-		}
-
-		// init Client
-		err = initClient(v)
-		if err != nil {
-			log.Default().Println("Error from init client:", err)
-			os.Exit(1)
-		}
-
-		// Check if time flag is set
-		if cmd.Flag(flagTime).Value.String() == "true" {
-			defer timeTrack(time.Now(), cmd.CommandPath())
-		}
-
-		// Get the vdc name from the command line
-		vdcName, err := cmd.Flags().GetString(flagName)
-		if err != nil {
-			fmt.Println("Malformed VDC name", err)
-			return
-		}
-
-		// Create the vdc
-		s.Stop()
-		fmt.Println("create vdc resource (with basic value)")
-		fmt.Println("vdc name: " + vdcName)
-		s.Restart()
-
-		if _, err = c.V1.VDC().New(&infrapi.CAVVirtualDataCenter{
-			VDC: infrapi.CAVVirtualDataCenterVDC{
-				Name:                vdcName,
-				ServiceClass:        "STD",
-				BillingModel:        "PAYG",
-				CPUAllocated:        22000,
-				VCPUInMhz:           2200,
-				Description:         "vdc created by cloudavenue-cli",
-				MemoryAllocated:     30,
-				DisponibilityClass:  "ONE-ROOM",
-				StorageBillingModel: "PAYG",
-				StorageProfiles: []infrapi.StorageProfile{
-					{
-						Class:   "gold",
-						Limit:   500,
-						Default: true,
-					},
-				},
-			},
-		}); err != nil {
-			fmt.Println("Error from vdc", err)
-			return
-		}
-		s.FinalMSG = "vdc resource created successfully !!\n"
-		s.Stop()
-	},
-}
-
-// createEdgeGatewayCmd create a edgegateway resource(s)
-var createEdgeGatewayCmd = &cobra.Command{
-	Use:               argEdgeGateway,
-	Short:             "Create an edgeGateway",
-	Aliases:           []string{"gw", "egw"},
-	Example:           "edgegateway create --vdc <vdc name> [--t0 <t0 name>]",
-	DisableAutoGenTag: true,
-	Run: func(cmd *cobra.Command, args []string) {
-		// init Config File
-		v, err := initConfig()
-		if err != nil {
-			log.Default().Println("Error from init config file", err)
-			return
-		}
-
-		// init Client
-		err = initClient(v)
-		if err != nil {
-			log.Default().Println("Error from init client:", err)
-			os.Exit(1)
-		}
-
-		// Check if time flag is set
-		if cmd.Flag(flagTime).Value.String() == "true" {
-			defer timeTrack(time.Now(), cmd.CommandPath())
-		}
-
-		// Get the vdc name from the command line
-		vdc, err := cmd.Flags().GetString("vdc")
-		if err != nil {
-			fmt.Println("Error from VDC", err)
-			return
-		}
-
-		// Get the t0 name
-		// if flag is not precise, get the first one
-		var t0 string
-		if cmd.Flag(argT0).Value.String() == "" {
-			t0s, err := c.V1.T0.GetT0s()
-			if err != nil {
-				fmt.Println("Error to retrieve your first T0", err)
-				return
-			}
-			t0 = (*t0s)[0].Tier0Vrf
-		} else {
-			t0, err = cmd.Flags().GetString("t0")
-			if err != nil {
-				fmt.Println("Error to retrieve T0", err)
-				return
-			}
-		}
-		// Create the edgeGateway
-		s.Stop()
-		fmt.Println("Creating EdgeGateway resource")
-		fmt.Println("vdc name: " + vdc)
-		fmt.Println("t0 name: " + t0)
-		s.Restart()
-		job, err := c.V1.EdgeGateway.New(vdc, t0)
-		if err != nil {
-			fmt.Println("Error from EdgeGateway", err)
-			return
-		}
-		err = job.Wait(3, 300)
-		if err != nil {
-			fmt.Println("Error during EdgeGateway Creation !!", err)
-			return
-		}
-		s.FinalMSG = "EdgeGateway resource created successfully !!"
-		s.Stop()
-	},
-}
-
-// createS3Cmd create a s3 bucket resource(s)
-var createS3Cmd = &cobra.Command{
-	Use:               argS3,
-	Short:             "Create an S3 bucket",
-	Example:           "create s3 --name <bucket name>",
-	DisableAutoGenTag: true,
-	Run: func(cmd *cobra.Command, args []string) {
-		// init Config File
-		v, err := initConfig()
-		if err != nil {
-			log.Default().Println("Error from init config file", err)
-			return
-		}
-
-		// init Client
-		err = initClient(v)
-		if err != nil {
-			log.Default().Println("Error from init client:", err)
-			os.Exit(1)
-		}
-
-		// Check if time flag is set
-		if cmd.Flag(flagTime).Value.String() == "true" {
-			defer timeTrack(time.Now(), cmd.CommandPath())
-		}
-
-		bucketName, err := cmd.Flags().GetString(flagName)
-		if err != nil {
-			fmt.Println("Malformed bucket name ", err)
-			return
-		}
-
-		// Create the bucket
-		s.Stop()
-		fmt.Println("create a bucket resource (with basic value)")
-		fmt.Println("bucket name: " + bucketName)
-		s.Restart()
-
-		_, err = c.V1.S3().CreateBucket(&s3.CreateBucketInput{Bucket: &bucketName})
-		if err != nil {
-			fmt.Println("Error from S3 Create", err)
-			return
-		}
-		s.FinalMSG = "Bucket resource created successfully !!"
-		s.Stop()
-	},
 }
