+++ cmd/del.go
@@ -1,14 +1,6 @@
 package cmd
 
 import (
-	"fmt"
-	"log"
-	"os"
-	"time"
-
-	"github.com/aws/aws-sdk-go/service/s3"
-	"github.com/orange-cloudavenue/cloudavenue-sdk-go/pkg/uuid"
-	v1 "github.com/orange-cloudavenue/cloudavenue-sdk-go/v1"
 	"github.com/spf13/cobra"
 )
 
@@ -58,214 +50,4 @@ func init() {
 	delCmd.AddCommand(delVDCCmd)
 	delVDCCmd.Args = cobra.MinimumNArgs(1)
 
-}
-
-// delVDCCmd delete a vdc resource(s)
-var delVDCCmd = &cobra.Command{
-	Use:               argVDC,
-	Example:           "del vdc <name> [<name>] [<name>] ...",
-	Short:             "Delete a vdc",
-	DisableAutoGenTag: true,
-	Run: func(cmd *cobra.Command, args []string) {
-		// init Config File
-		v, err := initConfig()
-		if err != nil {
-			log.Default().Println("Error from init config file", err)
-			return
-		}
-
-		// init Client
-		err = initClient(v)
-		if err != nil {
-			log.Default().Println("Error from init client:", err)
-			os.Exit(1)
-		}
-
-		// Check if time flag is set
-		if cmd.Flag(flagTime).Value.String() == "true" {
-			defer timeTrack(time.Now(), cmd.CommandPath())
-		}
-
-		for _, arg := range args {
-			s.Stop()
-			fmt.Println("delete vdc resource " + arg)
-			s.Restart()
-			vdc, err := c.V1.VDC().GetVDC(arg)
-			if err != nil {
-				fmt.Println("Error from vdc", err)
-				return
-			}
-			job, err := vdc.Delete()
-			if err != nil {
-				fmt.Println("Unable to delete vdc", err)
-				return
-			}
-			err = job.Wait(3, 300)
-			if err != nil {
-				fmt.Println("Error during vdc Deletion !!", err)
-				return
-			}
-			s.FinalMSG = "vdc resource deleted " + arg + " successfully !!\n"
-			s.Stop()
-		}
-
-	},
-}
-
-// deleteCmd delete a s3 bucket resource(s)
-var delS3Cmd = &cobra.Command{
-	Use:               argS3,
-	Example:           "delete s3 <name> [<name>] [<name>] ...",
-	Short:             "Delete a s3 bucket",
-	DisableAutoGenTag: true,
-	Run: func(cmd *cobra.Command, args []string) {
-		// init Config File
-		v, err := initConfig()
-		if err != nil {
-			log.Default().Println("Error from init config file", err)
-			return
-		}
-
-		// init Client
-		err = initClient(v)
-		if err != nil {
-			log.Default().Println("Error from init client:", err)
-			os.Exit(1)
-		}
-
-		// Check if time flag is set
-		if cmd.Flag(flagTime).Value.String() == "true" {
-			defer timeTrack(time.Now(), cmd.CommandPath())
-		}
-
-		for i, arg := range args {
-			s.Stop()
-			fmt.Println("delete bucket resource... " + arg)
-			s.Restart()
-			// Del the bucket
-			_, err := c.V1.S3().DeleteBucket(&s3.DeleteBucketInput{Bucket: &args[i]})
-			if err != nil {
-				fmt.Println("Error from S3 Delete", err)
-				return
-			}
-			s.FinalMSG = "Bucket resource deleted " + arg + " successfully !!\n"
-			s.Stop()
-		}
-
-	},
-}
-
-// deleteCmd delete a edgeGateway resource(s)
-var delEdgeGatewayCmd = &cobra.Command{
-	Use:               argEdgeGateway,
-	Aliases:           []string{argEdgeGatewayAlias1, argEdgeGatewayAlias2},
-	Example:           "delete edgegateway <id or name> [<id or name>] [<id or name>] ...",
-	Short:             "Delete an edgeGateway (name or id)",
-	DisableAutoGenTag: true,
-	Run: func(cmd *cobra.Command, args []string) {
-		// init Config File
-		v, err := initConfig()
-		if err != nil {
-			log.Default().Println("Error from init config file", err)
-			return
-		}
-
-		// init Client
-		err = initClient(v)
-		if err != nil {
-			log.Default().Println("Error from init client:", err)
-			os.Exit(1)
-		}
-
-		// Check if time flag is set
-		if cmd.Flag(flagTime).Value.String() == "true" {
-			defer timeTrack(time.Now(), cmd.CommandPath())
-		}
-		var (
-			gw *v1.EdgeGw
-		)
-
-		for _, arg := range args {
-			s.Stop()
-			fmt.Println("delete EdgeGateway resource " + arg)
-			s.Restart()
-			if uuid.IsUUIDV4(arg) {
-				gw, err = c.V1.EdgeGateway.GetByID(arg)
-			} else {
-				gw, err = c.V1.EdgeGateway.GetByName(arg)
-			}
-			if err != nil {
-				if err != nil {
-					fmt.Println("Unable to find EdgeGateway ID or Name", err)
-					return
-				}
-			}
-			job, err := gw.Delete()
-			if err != nil {
-				fmt.Println("Unable to delete EdgeGateway", err)
-				return
-			}
-			err = job.Wait(3, 300)
-			if err != nil {
-				fmt.Println("Error during EdgeGateway Deletion !!", err)
-				return
-			}
-			s.FinalMSG = "EdgeGateway resource deleted " + arg + " successfully !!\n"
-			s.Stop()
-		}
-
-	},
-}
-
-// deleteCmd delete a public ip resource(s)
-var delPublicIPCmd = &cobra.Command{
-	Use:               argPublicIP,
-	Aliases:           []string{argPublicIPAlias1},
-	Example:           "delete publicip <ip> [<ip>] [<ip>] ...",
-	Short:             "Delete public ip resource(s)",
-	DisableAutoGenTag: true,
-	Run: func(cmd *cobra.Command, args []string) {
-		// init Config File
-		v, err := initConfig()
-		if err != nil {
-			log.Default().Println("Error from init config file", err)
-			return
-		}
-
-		// init Client
-		err = initClient(v)
-		if err != nil {
-			log.Default().Println("Error from init client:", err)
-			os.Exit(1)
-		}
-
-		// Check if time flag is set
-		if cmd.Flag(flagTime).Value.String() == "true" {
-			defer timeTrack(time.Now(), cmd.CommandPath())
-		}
-
-		for _, arg := range args {
-			s.Stop()
-			fmt.Println("delete publicip resource " + arg)
-			s.Restart()
-			ip, err := c.V1.PublicIP.GetIP(arg)
-			if err != nil {
-				fmt.Println("Error from ip: ", err)
-				return
-			}
-			job, err := ip.Delete()
-			if err != nil {
-				fmt.Println("Unable to delete ip: ", err)
-				return
-			}
-			err = job.Wait(15, 300)
-			if err != nil {
-				fmt.Println("Error during ip Deletion !!", err)
-				return
-			}
-			s.FinalMSG = "ip resource deleted " + arg + " successfully !!\n"
-			s.Stop()
-		}
-
-	},
 }
