+++ cmd/get_vdc.go
@@ -1,0 +1,64 @@
+package cmd
+
+import (
+	"fmt"
+	"time"
+
+	"github.com/orange-cloudavenue/cloudavenue-cli/pkg/customErrors"
+	"github.com/orange-cloudavenue/cloudavenue-cli/pkg/output"
+	"github.com/orange-cloudavenue/common-go/print"
+	"github.com/spf13/cobra"
+)
+
+// getVDCCmd return a list of your vdc resource(s)
+var getVDCCmd = &cobra.Command{
+	Use:               argVDC,
+	Short:             "A brief list of your vdc resources",
+	Long:              "A complete list information of your s3 resources in your CloudAvenue account." + description,
+	DisableAutoGenTag: true,
+	SilenceErrors:     true,
+	RunE: func(cmd *cobra.Command, args []string) error {
+		// init Config File & Client
+		if err := initConfig(); err != nil {
+			return fmt.Errorf("Unable to initialize: %w", err)
+		}
+
+		// Check if time flag is set and print time elapsed
+		if cmd.Flag(flagTime).Value.String() == "true" {
+			defer timeTrack(time.Now(), cmd.CommandPath())
+		}
+
+		// Get the list of vdc
+		vdcs, err := c.V1.Querier().List().VDC()
+		if err != nil {
+			return fmt.Errorf("CloudAvenue Error from VDC List: %w", err)
+		}
+
+		// Print the result
+		s.Stop()
+		flag := cmd.Flag(flagOutput).Value.String()
+		w := print.New()
+		switch flag {
+		case flagOutputValueWide:
+			w.SetHeader("name", "status", "cpu used (mhz)", "memory used (mb)", "storage used (mb)", "number of vm(s)", "number of vapp(s)")
+			for _, v := range vdcs {
+				w.AddFields(v.Name, v.Status, *v.CpuUsedMhz, *v.MemoryUsedMB, *v.StorageUsedMB, *v.NumberOfVMs, *v.NumberOfVApps)
+			}
+		case flagOutputValueJSON, flagOutputValueYAML:
+			x, err := output.New(stringToTypeFormat(flag), vdcs)
+			if err != nil {
+				return fmt.Errorf("Impossible to format output: %w", err)
+			}
+			x.ToOutput()
+		case "":
+			w.SetHeader("name", "status")
+			for _, v := range vdcs {
+				w.AddFields(v.Name, v.Status)
+			}
+		default:
+			return fmt.Errorf("Output format %v: %w", flag, customErrors.ErrNotValidOutput)
+		}
+		w.PrintTable()
+		return nil
+	},
+}
